Refactor formatting code
