+++ po/POTFILES
@@ -62,3 +62,4 @@ src/DocumentationViewer.js
 src/langs/css/css.js
 
 src/langs/rust/Compiler.js
+src/langs/rust/rust.js
